MunitionSettings to its own file
